Restructured code to set proxy mode from command line
Application.main now checks for the first command line parameter containing the
substring "CGLIB" (case-insensitive). If found, it loads Spring configuration
class CglibPoxyConfig, otherwise defaults to JdkProxyConfig. The only difference
between the configurations is an the way @EnableAspectJAutoProxy is
parametrised: with or without 'proxyTargetClass = true'.

In order to make this work the config classes had to be moved to a package
outside the one configured for component scan, otherwise both configs would be
found. So now the application classes are in package de.scrum_master.spring.app
whereas the config classes reside in de.scrum_master.spring.config.